src/main/webapp/app/entities/accessSampleApplication/table-4/delete/table-4-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Table4Service } from '../service/table-4.service';

import { Table4DeleteDialogComponent } from './table-4-delete-dialog.component';

describe('Table4 Management Delete Component', () => {
  let comp: Table4DeleteDialogComponent;
  let fixture: ComponentFixture<Table4DeleteDialogComponent>;
  let service: Table4Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, Table4DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Table4DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Table4DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Table4Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
